Extract methods in JobCompletionData to Job.java
Moved methods and constructor from JobCompletionData to Job.java. Removed class JobCompletionData. Closes issue #4
