+++ src/applications/Job.java
@@ -2,18 +2,40 @@ package applications;
 
 import dataStructures.LinkedQueue;
 
-class Job {
+public class Job {
     // data members
     private LinkedQueue taskQ; // this job's tasks
     private int length; // sum of scheduled task times
     private int arrivalTime; // arrival time at current queue
     private int id; // job identifier
 
+    private  int completionTime;
+    private  int totalWaitTime;
+    private  int jobNumber;
     // constructor
-    Job(int theId) {
+       public Job(int theId) {
         id = theId;
         taskQ = new LinkedQueue();
         // length and arrivalTime have default value 0
+    }
+
+    // Second constructor that we will try to combine with first constructor in the future
+       public Job(int jobNumber, int completionTime, int totalWaitTime) {
+        this.jobNumber = jobNumber;
+        this.completionTime = completionTime;
+        this.totalWaitTime = totalWaitTime;
+    }
+
+   public int getCompletionTime() {
+        return completionTime;
+    }
+
+    public int getTotalWaitTime() {
+        return totalWaitTime;
+    }
+
+    public int getJobNumber() {
+        return jobNumber;
     }
 
     // other methods
